finished refractoring everything and got each component into its own file and working.
